Refactor DashboardLayout component to use a separate AppProviders component
