+++ src/app/dashboard/layout.tsx
@@ -1,28 +1,6 @@
-'use client'
-
 import React from 'react'
-import { ChakraProvider, createStandaloneToast } from '@chakra-ui/react'
-import { Provider as ReduxProviderClass } from 'react-redux'
-import { Dict } from '@chakra-ui/utils'
-import Fonts from '@/theme/Fonts'
-import { createConfig, WagmiConfig } from 'wagmi'
-import { connectors, publicClient, webSocketPublicClient } from '@/config/wagmi'
-import { store } from '@/store/store'
-import { UALProviderSwitch, WalletProvider } from '@/context/eosWalletContext'
-import { DashboardLayout } from '@/components/dashboard/layout'
-import GoogleAnalyticsScripts from '@/components/SEO/GoogleAnalyticsScripts'
-import chakraTheme from '@/theme'
 import { Metadata } from 'next'
-
-const { ToastContainer } = createStandaloneToast()
-const ReduxProvider = ReduxProviderClass as (props: Dict) => JSX.Element
-
-const client = createConfig({
-  autoConnect: true,
-  connectors,
-  publicClient,
-  webSocketPublicClient,
-})
+import AppProviders from '@/components/client/AppProviders'
 
 export const metadata: Metadata = {
   title: {
@@ -64,20 +42,7 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html lang="en">
       <body>
-        <ReduxProvider store={store}>
-          <ChakraProvider resetCSS theme={chakraTheme}>
-            <Fonts />
-            <WagmiConfig config={client}>
-              <GoogleAnalyticsScripts />
-              <UALProviderSwitch>
-                <WalletProvider>
-                  <DashboardLayout>{children}</DashboardLayout>
-                </WalletProvider>
-              </UALProviderSwitch>
-            </WagmiConfig>
-          </ChakraProvider>
-        </ReduxProvider>
-        <ToastContainer />
+        <AppProviders>{children}</AppProviders>
       </body>
     </html>
   )
